% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R, R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{distance}{distance in units}
  \item{damage}{damage caused}
  ...
}

A data frame with 15 rows and 2 variables:
\describe{
  \item{distance}{distance in units}
  \item{damage}{damage caused}
  ...
}
}
\source{
\url{Statistics for scientist and engineers}

\url{Statistics for scientist and engineers}
}
\usage{
fire

fire
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.

A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
