% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoeff.R
\name{getFunctionn}
\alias{getFunctionn}
\title{returns the value for the function given the values of x and the coefficients}
\usage{
getFunctionn(x, coef)
}
\arguments{
\item{x}{a data set}

\item{coef}{coefficients from summary}
}
\value{
the function generated
}
\description{
returns the value for the function given the values of x and the coefficients
}
