% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathMode.R
\name{mathMode}
\alias{mathMode}
\title{returns the mode of a vector}
\usage{
mathMode(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
the mode of a vector
}
\description{
returns the mode of a vector
}
