% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cltpois.R
\name{mycltp}
\alias{mycltp}
\title{Find the zscore}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{size of sample}

\item{iter}{number of samples to take}

\item{lambda}{poisson distribution parameter.}
}
\value{
histrograms of sample distributions and list of all samples taken.
}
\description{
Find the zscore
}
\examples{
mycltp(n=10, iter=10000, lambda=10)
}
