% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{function for the quadratic regression}
\usage{
myplot(x)
}
\arguments{
\item{x}{the breast height diameter}
}
\value{
Predicted the value from breast height diameter
}
\description{
function for the quadratic regression
}
