% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBinomial.R
\name{simBinomial}
\alias{simBinomial}
\title{Creates plots of the experiment described in parameters}
\usage{
simBinomial(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{# of experiments to run}

\item{n}{# of trials to run}

\item{p}{The probability of the distribution}
}
\description{
Creates plots of the experiment described in parameters
}
\examples{
simBinomial(100,10,0.5)

Runs 100 experiments with 10 trials. Symmetrically distributed round 5.
}
