% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBinomial.R
\name{simHyper}
\alias{simHyper}
\title{White and black balls in urn without replacement}
\usage{
simHyper(iter, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{# of experiments to run}

\item{N}{# of balls}

\item{r}{number of white balls}

\item{n}{# of balls drawn}
}
\description{
White and black balls in urn without replacement
}
\examples{
simHyper(iter=100,N=20, r=12, n=5)

Runs 100 experiments, selecting 5 balls without replacement each time. Plot is generated with frequency of white balls drawn.
}
