% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBinomial.R
\name{simSample}
\alias{simSample}
\title{White and black balls in urn without replacement}
\usage{
simSample(n, iter = 10, time = 0.5)
}
\description{
White and black balls in urn without replacement
}
\examples{
simSample(n=5,iter=100,time=0.5)

Run the experiment 100 times,
}
