% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore_func}
\alias{zscore_func}
\title{Find the zscore}
\usage{
zscore_func(mean, val, sd)
}
\arguments{
\item{mean}{the mean of a sample}

\item{val}{the value being examines}

\item{sd}{the standard deviation of the sample}
}
\value{
zscore of the value in a sample
}
\description{
Find the zscore
}
\examples{
zscore_func(89, 80, 5)
}
